Moved database.py to database module
